requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.74
langchain-openai==0.3.24
python-dotenv==1.0.0
wikipedia==1.4.0
google-search-results==2.4.2